+++ bin/build/tools/test/junit/hooks/testsuite-stop.sh
@@ -4,18 +4,16 @@
 #
 # Copyright &copy; 2026 Market Acumen, Inc.
 #
-# Hook: testsuite-end
+# Hook: testsuite-stop
 
 # shellcheck source=/dev/null
-if source "${BASH_SOURCE[0]%/*}/../../../tools.sh"; then
+if source "${BASH_SOURCE[0]%/*}/../../../../tools.sh"; then
 
-  # fn: hookRun testsuite-end
+  # fn: hookRun testsuite-stop
   # Summary: Run when a tests are started (before running)
   # Argument: stateFile - File. Required. State file for test suite.
   __hookTestSuiteStop() {
     local handler="_${FUNCNAME[0]}"
-
-    local home && home=$(catchReturn "$handler" buildHome) || return $?
 
     local suiteName="" stateFile=""
 
@@ -27,10 +25,10 @@ if source "${BASH_SOURCE[0]%/*}/../../../tools.sh"; then
       [ -n "$argument" ] || throwArgument "$handler" "blank #$__index/$__count ($(decorate each quote -- "${__saved[@]}"))" || return $?
       case "$argument" in
       *)
-        if [ -z "$stateFile" ]; then
-          suiteName=$(validate "$handler" String "$argument") || return $?
+        if [ -z "$suiteName" ]; then
+          suiteName=$(validate "$handler" String "suiteName" "$argument") || return $?
         elif [ -z "$stateFile" ]; then
-          stateFile=$(validate "$handler" File "$argument") || return $?
+          stateFile=$(validate "$handler" File "stateFile" "$argument") || return $?
         fi
         ;;
       esac
@@ -40,19 +38,18 @@ if source "${BASH_SOURCE[0]%/*}/../../../tools.sh"; then
     [ -n "$suiteName" ] || throwArgument "$handler" "suiteName is required" || return $?
     [ -n "$stateFile" ] || throwArgument "$handler" "stateFile is required" || return $?
 
+    local suiteName="$TEST_SUITE_NAME"
     local returnCode=0 && (
-      local undo=(rm -rf "$junitSuiteTemp")
-
-      catchReturn "$handler" source "$stateFile" || returnUndo $? "${undo[@]}" || return $?
+      local junitTemp="" tests=0 failures=0 errors=0 skipped=0 testFile="" junitSuiteTemp="" startSuite=0 junitKeepTemp=false timestampSuite=""
+      __testLoader "$handler" : || return $?
+      catchReturn "$handler" source "$stateFile" || return $?
       if [ -d "$junitSuiteTemp" ]; then
-        local suiteXML="$junitTemp/.xml/$suiteName.xml"
-        local junitTemp="" suiteName="" tests=0 failures=0 errors=0 skipped=0 testFile="" junitSuiteTemp="" startSuite=0
-        local timestampSuite=""
+        local suiteXML="$junitTemp/.suiteXML/$suiteName.xml" suiteStats="$junitTemp/$suiteName.stats"
 
-        local suiteStats="$junitTemp/$suiteName.stats"
-        __collectStats "$handler" "$junitSuiteTemp" tests failures errors skipped >"$suiteStats" || return $?
+        __collectStats "$handler" "$junitSuiteTemp" tests success failures errors skipped >"$suiteStats" || return $?
         catchReturn "$handler" source "$suiteStats" || return $?
-        catchReturn "$handler" rm -f "$suiteStats" || return $?
+        __mergeStats "$handler" "$junitSuiteTemp" "$junitTemp/.totals/" tests success failures errors skipped || return $?
+        $junitKeepTemp || catchReturn "$handler" rm -f "$suiteStats" || return $?
 
         {
           local aa=(
@@ -61,24 +58,28 @@ if source "${BASH_SOURCE[0]%/*}/../../../tools.sh"; then
             "failures=$failures"
             "errors=$errors"
             "skipped=$skipped"
-            "time=$(($(timingElapsed "$startSuite") / 1000))"
+            "time=$(timingFormat "$(timingElapsed "$startSuite")")"
             "file=$testFile"
             "timestamp=$timestampSuite"
           )
-          catchReturn "$handler" junitSuiteOpen "${aa[@]}" || returnUndo $? "${undo[@]}" || return $?
-          catchReturn "$handler" find "$junitSuiteTemp" -type f | sort -u | xargs cat || returnUndo $? "${undo[@]}" || return $?
+          catchReturn "$handler" junitSuiteOpen "${aa[@]}" || return $?
+          catchReturn "$handler" find "$junitSuiteTemp" -type f -name '*.xml' | sort -u | xargs cat || return $?
           catchReturn "$handler" junitSuiteClose || return $?
-        } >>"$suiteXML" || returnUndo $? "${undo[@]}" || return $?
-        catchReturn "$handler" rm -rf "$junitSuiteTemp" || return $?
+        } >>"$suiteXML" || return $?
+        $junitKeepTemp || catchReturn "$handler" rm -rf "$junitSuiteTemp" || return $?
         catchReturn "$handler" environmentValueWrite junitSuiteTemp "" >>"$stateFile" || return $?
       fi
     ) || returnCode=$?
-    catchReturn "$handler" hookRunOptional --application "$home" --next "${BASH_SOURCE[0]}" "$HOOK_NAME" "$@" || return $?
+
+    # IDENTICAL hookRunOptionalNext 2
+    local home && home=$(catchReturn "$handler" buildHome) || return $?
+    catchReturn "$handler" hookRunOptional --application "$home" --next "${BASH_SOURCE[0]}" "$HOOK_NAME" "${__saved[@]+"${__saved[@]}"}" || return $?
+
     return "$returnCode"
   }
   ___hookTestSuiteStop() {
     # __IDENTICAL__ usageDocument 1
     usageDocument "${BASH_SOURCE[0]}" "${FUNCNAME[0]#_}" "$@"
   }
-  __hookTestSuiteEnd "$@"
+  __hookTestSuiteStop "$@"
 fi
